% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i_dummy_cols.R
\name{i_dummy_cols}
\alias{i_dummy_cols}
\title{Fast creation of dummy variables}
\usage{
i_dummy_cols(
  .data,
  select_columns = NULL,
  remove_first_dummy = FALSE,
  remove_most_frequent_dummy = FALSE,
  ignore_na = FALSE,
  split = NULL,
  remove_selected_columns = FALSE,
  omit_colname_prefix = FALSE,
  sep = "_"
)
}
\arguments{
\item{.data}{An object with the data set you want to make dummy columns from.}

\item{select_columns}{Vector of column names that you want to create dummy variables from.
If NULL (default), uses all character and factor columns.}

\item{remove_first_dummy}{Removes the first dummy of every variable such that only n-1 dummies remain.
This avoids multicollinearity issues in models.}

\item{remove_most_frequent_dummy}{Removes the most frequently observed category such that only n-1 dummies
remain. If there is a tie for most frequent, will remove the first
(by alphabetical order) category that is tied for most frequent.}

\item{ignore_na}{If TRUE, ignores any NA values in the column. If FALSE (default), then it
will make a dummy column for value_NA and give a 1 in any row which has a
NA value.}

\item{split}{A string to split a column when multiple categories are in the cell. For
example, if a variable is Pets and the rows are "cat", "dog", and "turtle",
each of these pets would become its own dummy column. If one row is "cat, dog",
then a split value of "," this row would have a value of 1 for both the cat
and dog dummy columns.}

\item{remove_selected_columns}{If TRUE (not default), removes the columns used to generate the dummy columns.}

\item{omit_colname_prefix}{If TRUE (not default) and \code{length(select_columns) == 1}, omit pre-pending the
name of \code{select_columns} to the names of the newly generated dummy columns}

\item{sep}{A string to use as separator between column name and unique value.}
}
\value{
A data.frame (or tibble or data.table, depending on input data type) with
same number of rows as inputted data and original columns plus the newly
created dummy columns.
}
\description{
Quickly create dummy (binary) columns from character and
factor type columns in the inputted data (and numeric columns if specified.)
This function is useful for statistical analysis when you want binary
columns rather than character columns.
}
\examples{
crime <- data.frame(city = c("SF", "SF", "NYC"),
                    year = c(1990, 2000, 1990),
                    crime = 1:3)
i_dummy_cols(crime, sep = "_-_")
}
\concept{dummy functions}
